Another reorg. Fix lshaped -> lshapes in class.
