+++ src/editor/codemirror/structure-highlighting/index.ts
@@ -1,15 +1,40 @@
 import { Compartment, Extension } from "@codemirror/state";
 import { CodeStructureHighlight } from "../../../settings/settings";
-import { codeStructure } from "./analysis";
+import { baseTheme, themeTweakForBackgrounds } from "./theme";
+import { codeStructureView } from "./view";
 
 export const structureHighlightingCompartment = new Compartment();
+
+export interface CodeStructureSettings {
+  shape: "l-shape" | "box";
+  background: "block" | "none";
+  borders: "borders" | "no-borders" | "left-edge-only";
+
+  hoverBackground?: boolean;
+  cursorBackground?: boolean;
+  hoverBorder?: boolean;
+  cursorBorder?: boolean;
+}
+
+// We'll switch to exporting this soon, see below.
+const codeStructure = (settings: CodeStructureSettings): Extension => [
+  codeStructureView(settings),
+  baseTheme,
+  settings.background !== "none" ||
+  settings.cursorBackground ||
+  settings.hoverBackground
+    ? themeTweakForBackgrounds
+    : [],
+];
+
+// This will go soon in favour of more fine-grained settings
 export const structureHighlighting = (
   option: CodeStructureHighlight
 ): Extension => {
   switch (option) {
     case "l-shapes":
       return codeStructure({
-        shape: "box",
+        shape: "l-shape",
         background: "none",
         borders: "left-edge-only",
       });
